Fixed the app error of the LED button press animation and the functionality of deleting the bonds and restarting advertising

+++ Proximo_nRF/Peripheral/io.h
@@ -14,7 +14,7 @@
   void proximo_tps_on   (void);
   void proximo_tps_off  (void);
   uint32_t proximo_tps_read_output(void);
-  void enable_tps (uint8_t ms_delay);
+  void enable_tps (uint16_t ms_delay);
 
   void proximo_ldr_on   (void);
   void proximo_ldr_off  (void);
